Semana 9 - v2

+++ proyecto/src/main/java/com/proyecto/repository/CategoriasRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface CategoriasRepository extends CrudRepository<Categorias,Long>{
+public interface CategoriasRepository extends CrudRepository<Categorias,Integer>{
 
 }
