+++ proyecto/src/main/java/com/proyecto/repository/UrlRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface UrlRepository extends CrudRepository<Url,Long>{
+public interface UrlRepository extends CrudRepository<Url,Integer>{
 
 }
